refactor(blazemeter): colocate shared runtime
