+++ src/providers/blaze_meter_functional/actions.ts
@@ -2,7 +2,7 @@ import type { ActionDefinition } from "../../core/types.ts";
 
 import { s } from "../../core/json-schema.ts";
 import { defineProviderAction } from "../../core/provider-definition.ts";
-import { blazeMeterResponseEnvelopeSchema, blazeMeterSortSchema } from "../blaze-meter-schemas.ts";
+import { blazeMeterResponseEnvelopeSchema, blazeMeterSortSchema } from "../blaze_meter_performance/shared-schemas.ts";
 
 const service = "blaze_meter_functional";
